+++ tests/text_to_ast/text_only.rs
@@ -609,3 +609,31 @@ fn link_in_parenthesis2() {
         ]
     );
 }
+
+
+#[test]
+fn bot_suggestion_is_no_email() {
+    assert_eq!(
+        parse_only_text("/command@bot@addr.com"),
+        vec![
+            BotCommandSuggestion("/command@bot@addr.com"),
+        ]
+    );
+    assert_eq!(
+        parse_only_text("\n/command@bot@addr.com"),
+        vec![
+            Linebreak,
+            BotCommandSuggestion("/command@bot@addr.com"),
+        ]
+    );
+
+    assert_eq!(
+        parse_only_text("Bots that can be selected \n/command@bot@addr.com BOT"),
+        vec![
+            Text("Bots that can be selected "),
+            Linebreak,
+            BotCommandSuggestion("/command@bot@addr.com"),
+            Text(" BOT"),
+        ]
+    );
+}